HDDS-15520. Reduce duplication in TestOzoneManagerHAFollowerRead
